refactor: split apart the modern and legacy components
